pyproject.toml
[tool.poetry]
name = "pl_squad"
version = "0.1"
description = ""
authors = ["Amadeusz Oleszczak <aoleszczak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11.13"
loguru = "^0.7.2"
numpy = "^1.26.4"
poethepoet = "0.34.0"
torch = "2.6.0"
pandas = "2.2.3"
pyrate-limiter = "3.7.1"
python-dotenv = "^1.0.1"

# Feature engineering
langchain = "^0.3.25"
sentence-transformers = "^3.0.0"

# Agents
langchain-openai = "^0.3.24"
langchain-community = "^0.3.25"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"
pre-commit = "^3.7.1"
pytest = "^8.2.2"
isort = "^5.0.0"


# ----------------------------------
# --- Poe the Poet Configuration ---
# ----------------------------------
ipykernel = "^6.29.5"

[tool.poe.tasks]
# QA
lint-check = "poetry run ruff check ."
format-check = "poetry run ruff format --check ."

[tool.isort]
profile = "black"
